Enable four mechanical ty rules
Fix all diagnostics for not-iterable, call-non-callable,
unsupported-operator, and invalid-assignment, and remove their ignore
lines from [tool.ty.rules]. Highlights:

- ArrayLike is now Annotated[np.ndarray, ...] instead of object, which
  also removes many diagnostics under the still-ignored rules.
- is_array/is_pathlike gained TypeGuard/TypeIs so their checks narrow.
- get_quantity/get_factor_and_unit use plain unions instead of the
  misused str_or_none TypeVar; invert_quantity now returns None instead
  of raising TypeError when given empty units.
- get_array and BaseCoord.__getitem__ annotations no longer claim coord
  returns for array/scalar results.
- TDMS unsupported-data-type error no longer raises TypeError for
  unknown type codes; spool.map guards os.cpu_count() returning None.
- Refresh the burn-down counts comment in pyproject.toml.

+++ dascore/transform/fbe.py
@@ -84,7 +84,8 @@ def fbe(
     check_filter_range(nyquist, low, high, filt_min, filt_max)
 
     if step is None:
-        step = to_float(1 / sample_rate)
+        # to_float's scalar path returns a float despite its ndarray hint.
+        step = float(to_float(1 / sample_rate))
 
     patch = patch.pass_filter(**kwargs)
